Odik 28 item additional

+++ src/main/java/com/micutne/odik/service/TourItemService.java
@@ -1,5 +1,6 @@
 package com.micutne.odik.service;
 
+import ch.qos.logback.core.joran.action.PreconditionValidator;
 import com.micutne.odik.common.exception.AuthException;
 import com.micutne.odik.common.exception.BusinessException;
 import com.micutne.odik.common.exception.ErrorCode;
@@ -18,6 +19,7 @@ import com.micutne.odik.utils.file.Extensions;
 import jakarta.persistence.criteria.Predicate;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.mapstruct.control.MappingControl;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
@@ -48,51 +50,58 @@ public class TourItemService {
      */
 
     public TourItemResponse readOne(String reference_id) {
-        TourItem tourItem = tourItemRepository.findByReferenceIdGoogle(reference_id);
-        return TourItemResponse.fromEntity(tourItemRepository.findByReferenceIdGoogle(reference_id));
+        if (tourItemRepository.existsByReferenceIdGoogleAndState(reference_id, "public")) {
+            TourItem tourItem = tourItemRepository.findByReferenceIdGoogle(reference_id);
+            return TourItemResponse.fromEntity(tourItemRepository.findByReferenceIdGoogle(reference_id));
+        }
+        return TourItemResponse.resultMessage("TOUR_ITEM_NOT_EXIST");
     }
 
     /**
      * 관광지 저장
      */
     @Transactional
-    public TourItemResponse create(TourItemRequest request, String id) {
+    public TourItemResponse create(TourItemRequest request, String username) {
+
         if (tourItemRepository.existsByReferenceIdGoogle(request.getReference_id_google()))
             return new TourItemResponse(String.valueOf(ErrorCode.TOUR_ITEM_ALREADY_EXIST));
 
-        request.setUser(userRepository.findByIdOrThrow(id));
-        TourItem tourItem = tourItemMapper.toEntity(request);
-        tourItem = tourItemRepository.save(tourItem);
+        if (userRepository.existsById(username)) {
+            request.setUser(userRepository.findByIdOrThrow(username));
+            TourItem tourItem = tourItemMapper.toEntity(request);
+            tourItem = tourItemRepository.save(tourItem);
 
-        List<ImageTourItem> imageTourItems = new ArrayList<>();
+            List<ImageTourItem> imageTourItems = new ArrayList<>();
 
-        if (request.getImages_google() != null) {
-            for (String imageUrl : request.getImages_google()) {
-                // Url 또는 파일 추가 로직 ( 파일 추가부분 수정 필요 )
-                boolean saveUrl = imageUrl.startsWith("http://") || imageUrl.startsWith("https://");
-                String saveFile = "saveFile";
-                ImageTourItem imageTourItem = ImageTourItem.builder()
-                        .tour_item_idx(tourItem)
-                        .url(saveUrl ? imageUrl : saveFile)
-                        .build();
-                imageTourItems.add(imageTourItem);
+            if (request.getImages_google() != null) {
+                for (String imageUrl : request.getImages_google()) {
+                    // Url 또는 파일 추가 로직 ( 파일 추가부분 수정 필요 )
+                    boolean saveUrl = imageUrl.startsWith("http://") || imageUrl.startsWith("https://");
+                    String saveFile = "saveFile";
+                    ImageTourItem imageTourItem = ImageTourItem.builder()
+                            .tour_item_idx(tourItem)
+                            .url(saveUrl ? imageUrl : saveFile)
+                            .build();
+                    imageTourItems.add(imageTourItem);
+                }
             }
+
+            imageTourItems = imageTourItemRepository.saveAll(imageTourItems);
+
+            TourItemResponse response = TourItemResponse.fromEntity(tourItem, imageTourItems);
+            response.setImages_google(imageTourItems.stream().map(ImageTourItem::getImagesGoogle).toList());
+            return response;
         }
-
-        imageTourItems = imageTourItemRepository.saveAll(imageTourItems);
-
-        TourItemResponse response = TourItemResponse.fromEntity(tourItem, imageTourItems);
-        response.setImages_google(imageTourItems.stream().map(ImageTourItem::getImagesGoogle).toList());
-        return response;
+        return TourItemResponse.resultMessage("AUTH_FAIL");
     }
 
     /**
      * 관광지 수정
      */
     @Transactional
-    public TourItemResponse update(int idx, TourItemRequest request, String id) {
+    public TourItemResponse update(int idx, TourItemRequest request, String username) {
         TourItem tourItem = tourItemRepository.findByIdOrThrow(idx);
-        User user = userRepository.findByIdOrThrow(id);
+        User user = userRepository.findByIdOrThrow(username);
         checkAuth(tourItem, user);
         tourItem.updateTourItem(request);
         return TourItemResponse.fromEntity(tourItem);
@@ -102,16 +111,17 @@ public class TourItemService {
     /**
      * 관광지 삭제
      */
-    public void remove(int idx, String id) {
-        TourItem tourItem = tourItemRepository.findByIdOrThrow(idx);
-        User user = userRepository.findByIdOrThrow(id);
-        checkAuth(tourItem, user);
-        try {
+    public TourItemResponse remove(int idx, String username) {
+        User user = userRepository.findByIdOrThrow(username);
+
+        if (tourItemRepository.existsByIdx(idx)) {
+            TourItem tourItem = tourItemRepository.findByIdOrThrow(idx);
+            checkAuth(tourItem, user);
+
             tourItemRepository.delete(tourItem);
-        } catch (Exception e) {
-            throw new BusinessException(ErrorCode.TOUR_ITEM_DELETE_FAIL);
+            return TourItemResponse.resultMessage("OK");
         }
-        log.info("delete a tour item");
+        return TourItemResponse.resultMessage("TOUR_ITEM_NOT_EXIST");
     }
 
     /**
@@ -169,9 +179,8 @@ public class TourItemService {
     /**
      * 사용자 확인(추후 수정 및 추가)
      */
+
     public void checkAuth(TourItem tourItem, User user) {
         if (!tourItem.getUser().equals(user)) throw new AuthException(ErrorCode.USER_NOT_FOUND);
     }
-
-
 }
